Be more robust to errors
Bubble errors up from DiscordClient through to background/index,
ultimately showing a red badge on the action icon if the port could not
connect.

Also splits out connectNative from the DiscordClient constructor.  This
allows the synchronous/lastError to be checked, and making the validity
of port within DiscordClient easier to reason about.

Also fixes a bug with incorrect Promise chaining, and cleans up the way
async and Promises are used in some places.