Make content script persistent
Instead of injecting a single function every time the service worker
wants to query the activity, inject the code once and then send it a
message.

This makes testing simpler since filterMatches and getActivity can be
unit-tested.

It also paves the way for the content script to observe mutations in the
DOM.