Fix badge status reporting
Correctly catch exceptions thrown from the Discord client (fixes
uncaught exception in the console, and correctly sets warning badge
again).

Also change promise chaining logic in ActivityManager.setActivity so
that there is 1 promise added to the chain per call and calls to
maybeConnectAndSend are fully serialized.

Also add the state to the tooltip and consistently end the tooltips in a
period.